+++ packages/interaction-kit/bin/ikit.ts
@@ -1,10 +1,9 @@
 #!/usr/bin/env node
 import dotenv from "dotenv";
 import arg from "arg";
-import pkg from "../package.json";
+import pkg from "../package.json" assert { type: "json" };
 import devCommand from "../src/cli/dev";
 import deployCommand from "../src/cli/deploy";
-import serverCommand from "../src/cli/server";
 
 // Credit to Next.js, which I largely ripped off for this
 
@@ -20,7 +19,6 @@ process.on("SIGINT", () => process.exit(0));
 const commands: Record<string, (argv?: string[]) => void> = {
 	dev: devCommand,
 	deploy: deployCommand,
-	server: serverCommand,
 };
 
 const args = arg(
