+++ packages/interaction-kit/package.json
@@ -32,7 +32,7 @@
 		"arg": "^5.0.1",
 		"chokidar": "^3.5.2",
 		"cross-spawn": "^7.0.3",
-		"discord-api-types": "0.28.0-next.3503a4f.1646275715",
+		"discord-api-types": "0.29.0",
 		"dotenv": "^10.0.0",
 		"fastify": "^3.25.0",
 		"fastify-raw-body": "^3.1.0",
